# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  dtos.ActionDTO:
    properties:
      code:
        type: string
      device_id:
        type: string
      remote_id:
        type: string
      topic:
        type: string
      value: {}
    type: object
...
  dtos.TranscriptionTaskResponseDTO:
    properties:
      recording_id:
        example: uuid-v4
        type: string
      task_id:
        example: abc-123
        type: string
      task_status:
        example: pending
        type: string
    type: object
...
  dtos.UpdateSceneRequestDTO:
    properties:
      actions:
        items:
          $ref: '#/definitions/dtos.ActionDTO'
        type: array
      name:
        type: string
    required:
    - name
    type: object
  recordings_dtos.GetAllRecordingsResponseDto:
    properties:
      limit:
        type: integer
      page:
        type: integer
      recordings:
        items:
          $ref: '#/definitions/recordings_dtos.RecordingResponseDto'
        type: array
      total:
        type: integer
    type: object
  recordings_dtos.RecordingResponseDto:
    properties:
      audio_url:
        type: string
      created_at:
        type: string
      filename:
        type: string
      id:
        type: string
      original_name:
        type: string
    type: object
  recordings_dtos.StandardResponse:
    properties:
      data: {}
      details:
        description: |-
          Details is only populated for 400 (Bad Request) and 422 (Unprocessable Entity) errors.
          For all other status codes, including 500, this field is nil/omitted.
      message:
        example: Success
        type: string
      status:
        example: true
        type: boolean
    type: object
  sensio_domain_common_dtos.StandardResponse:
    properties:
      data: {}
      details:
        description: |-
          Details is only populated for 400 (Bad Request) and 422 (Unprocessable Entity) errors.
          For all other status codes, including 500, this field is nil/omitted.
      message:
        example: Success
        type: string
      status:
        example: true
        type: boolean
    type: object
...
  sensio_domain_speech_dtos.StandardResponse:
    properties:
      data: {}
      details: {}
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for Sensio App. <br> For full documentation,
    visit <a href="/docs">/docs</a>.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sensio API
  version: "1.0"
paths:
...
  /api/devices/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteDevice
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: DeleteDevice
      tags:
      - 09. Terminals
    get:
      consumes:
      - application/json
      description: GetDeviceByID
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: GetDeviceByID
      tags:
      - 09. Terminals
    put:
      consumes:
      - application/json
      description: UpdateDevice
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: UpdateDevice
      tags:
      - 09. Terminals
...
  /api/recordings:
    get:
      description: Get a paginated list of all recordings.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/recordings_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/recordings_dtos.GetAllRecordingsResponseDto'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: List all recordings
      tags:
      - 07. Recordings
    post:
      consumes:
      - multipart/form-data
      description: Upload an audio file and save its metadata.
      parameters:
      - description: Audio file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/recordings_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/recordings_dtos.RecordingResponseDto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/recordings_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Save a new recording
      tags:
      - 07. Recordings
...
  /api/speech/models/groq:
    post:
      consumes:
      - multipart/form-data
      description: Submit audio file for transcription via Groq. Processing is asynchronous.
      parameters:
      - description: Audio file (.mp3, .wav, .m4a, .aac, .ogg, .flac)
        in: formData
        name: audio
        required: true
        type: file
      - description: Language code (e.g. id, en)
        in: formData
        name: language
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.TranscriptionTaskResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Transcribe audio file (Groq)
      tags:
      - 04. Speech
...
  /api/speech/models/whisper/cpp:
    post:
      consumes:
      - multipart/form-data
      description: Submit audio file for transcription via Whisper.cpp. Processing
        is asynchronous.
      parameters:
      - description: Audio file (.mp3, .wav, .m4a, .aac, .ogg, .flac)
        in: formData
        name: audio
        required: true
        type: file
      - description: Language code (e.g. id, en)
        in: formData
        name: language
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dtos.TranscriptionTaskResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sensio_domain_speech_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Transcribe audio file (Whisper.cpp)
      tags:
      - 04. Speech
...
  /api/terminal/{id}/scenes/{scene_id}:
    delete:
      description: Remove a specific scene configuration
      parameters:
      - description: Terminal UUID
        in: path
        name: id
        required: true
        type: string
      - description: Scene UUID
        in: path
        name: scene_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Scene deleted
          schema:
            $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete a scene
      tags:
      - 03. Scenes
    put:
      consumes:
      - application/json
      description: Update the configuration of a specific scene
      parameters:
      - description: Terminal UUID
        in: path
        name: id
        required: true
        type: string
      - description: Scene UUID
        in: path
        name: scene_id
        required: true
        type: string
      - description: Updated scene configuration
        in: body
        name: scene
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateSceneRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Scene updated
          schema:
            $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
        "404":
          description: Scene not found
          schema:
            $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update an existing scene
      tags:
      - 03. Scenes
  /api/terminal/{id}/scenes/{scene_id}/control:
    get:
      description: Trigger all actions defined in a specific scene
      parameters:
      - description: Terminal UUID
        in: path
        name: id
        required: true
        type: string
      - description: Scene UUID
        in: path
        name: scene_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Scene applied
          schema:
            $ref: '#/definitions/sensio_domain_common_dtos.StandardResponse'
      security:
      - BearerAuth: []
      summary: Apply/Trigger a scene
      tags:
      - 03. Scenes
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...