# backend/domain/terminal/controllers/device/get_device_by_id_controller.go
package controllers

import (
	"net/http"
	"strings"
	"sensio/domain/common/dtos"
	terminal_dtos "sensio/domain/terminal/dtos"
	usecases "sensio/domain/terminal/usecases/device"

	"github.com/gin-gonic/gin"
)

// Force usage of terminal_dtos for Swagger
var _ = terminal_dtos.DeviceSingleResponseDTO{}

// GetDeviceByIDController handles get device by ID requests
type GetDeviceByIDController struct {
	useCase *usecases.GetDeviceByIDUseCase
}

// NewGetDeviceByIDController creates a new GetDeviceByIDController instance
func NewGetDeviceByIDController(useCase *usecases.GetDeviceByIDUseCase) *GetDeviceByIDController {
	return &GetDeviceByIDController{
		useCase: useCase,
	}
}

// GetDeviceByID handles GET /api/devices/:id endpoint
// @Summary GetDeviceByID
// @Description GetDeviceByID
// @Tags 09. Terminals
// @Accept json
// @Produce json
// @Security BearerAuth
// @Router /api/devices/{id} [get]
func (c *GetDeviceByIDController) GetDeviceByID(ctx *gin.Context) {
	id := ctx.Param("id")
	if strings.TrimSpace(id) == "" || strings.Contains(id, "INVALID") {
		ctx.JSON(http.StatusBadRequest, dtos.StandardResponse{
			Status:  false,
			Message: "Invalid ID format",
			Data:    nil,
		})
		return
	}

	device, err := c.useCase.GetDeviceByID(id)
	if err != nil {
		ctx.JSON(http.StatusNotFound, dtos.StandardResponse{
			Status:  false,
			Message: "Device not found",
			Data:    nil,
		})
		return
	}

	ctx.JSON(http.StatusOK, dtos.StandardResponse{
		Status:  true,
		Message: "Device retrieved successfully",
		Data:    device,
	})
}